Evitar warnings de compilación
@CatOfTheCannals, por qué hay código comentado?

+++ data/grafGen.cpp
@@ -26,7 +26,7 @@ void genKGraph(int n, edgeList &edges, int maxW, bool digraph){
 void shuffleEdgeList(edgeList &edges) { //https://codeground.in/blog/index.php/2016/11/13/technical-interview-question-on-data-structures-and-algorithms-perfect-shuffle/
     unsigned seed = std::chrono::system_clock::now().time_since_epoch().count();
     std::default_random_engine generator (seed);
-    for (int i = 0; i < edges.size(); ++i) {
+    for (size_t i = 0; i < edges.size(); ++i) {
         std::uniform_int_distribution<int> distribution(i, edges.size()-1);
         std::swap(edges[i], edges[distribution(generator)]);
     }
